fix/update query logic for like,unlike, get likes

+++ routes/listRemoveLike.js
@@ -4,21 +4,33 @@ const { Pool, Client } = require('pg');
 var db = require('../db');
 
 /* GET users listing. */
-router.delete('/', function(req, res, next) {
-  db.query(`
-    DELETE FROM 
-      likes
-    WHERE list_id = $1
-    AND user_id = $2
-    `, [req.body.list_id, req.body.user_id], (err, result)=>{
-    if(err) {
-      console.log(err)
-      return next(err);
-    } else {
-      data = result;
-      res.json(data.rows);
+router.delete('/', async function(req, res, next) {
+
+  const query1 = {
+    text: `DELETE FROM likes WHERE list_id = $1 AND user_id = $2`,
+    values: [req.body.list_id, req.body.user_id]
+  }
+
+  const query2 = {
+    text: `SELECT COUNT(*) FROM LIKES where list_id = $1`,
+    values: [req.body.list_id]
+  }
+
+  const q1 = await db.query(query1);
+  const q2 = await db.query(query2);
+
+  console.log('**************************************')
+  console.log('DELETE LIKE');
+  console.log('user id: ',req.body.user_id);
+  console.log('list id: ',req.body.list_id);
+
+  res.json(
+    {
+      count: q2.rows[0].count,
+      voted: q1.rows.length > 0 ? true : false
     }
-  });
+  );
 });
 
 module.exports = router;
+
